Logic with gets and chomps implemented on the main file
