+++ bin/main.rb
@@ -1,3 +1,23 @@
+#!/usr/bin/env ruby
+
+require_relative '../lib/board.rb'
+
+# Game starter
+
+def start_game
+  starter = true
+  game = Game.new
+  count = 1
+  while starter
+    game.new_board unless count == 1
+    game.new_player unless count == 1
+    game.game
+    starter = game.new_game
+    count += 1
+  end
+end
+
+# Game setup
 
 class Game
   attr_reader :player1
@@ -88,5 +108,39 @@ class Game
       puts option != 'Y' && option != 'N' ? "Please select Y or N as your option" : option != 'N'? "Ok! let's start a new game" : 'Thanks for playing!'
     end
     return option == 'Y' ? true : false
+  end
+end
+
+# Player settings 
+
+class Player
+  attr_reader :symbol
+  @@available_symbol = ['X', 'O']
+  @@remaining_symbol = Array.new
+  def initialize
+    @symbol = player_id(nil)
+  end
+  def player_id (symbol)
+    if @@available_symbol.size > 1
+      puts 'Select your symbol, either X or O'
+      while symbol != 'X' && symbol != 'O'
+        symbol = gets.chomp
+        symbol.upcase!
+        puts symbol != 'X' && symbol != 'O' ? "Please select X or O as your symbol" : "Cool! #{symbol} choosen"
+      end
+      puts @@remaining_symbol = @@available_symbol.reject!{|x| x == symbol}
+      symbol
+    elsif @@available_symbol.size > 0
+      puts "next player gets the #{@@remaining_symbol[0]} symbol"
+      symbol = @@remaining_symbol[0]
+      @@remaining_symbol = @@available_symbol.reject!{|x| x == symbol}
+      symbol
+    else
+      puts 'No more players can be created'
+    end
+  end
+  def restart
+    @@available_symbol = ['X', 'O']
+    @@remaining_symbol = Array.new
   end
 end
